+++ activerecord/test/cases/helper.rb
@@ -74,6 +74,7 @@ end
   supports_insert_conflict_target?
   supports_optimizer_hints?
   supports_datetime_with_precision?
+  supports_use_index_hints?
 ].each do |method_name|
   define_method method_name do
     ActiveRecord::Base.connection.public_send(method_name)